requirements.txt
aiogram==3.6.0
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary==2.9.9
gtts==2.6.0
SpeechRecognition==3.10.0
pydub==0.25.1